Refactor setup to have better architecture
