+++ test/ai-development/evals/suites/testing-skill-routing/artifact-grader.mjs
@@ -1,94 +1,19 @@
 /**
- * Artifact capture and grading for this suite.
+ * Deterministic artifact grading for this suite.
  */
-import { execFile } from 'node:child_process';
-import { promisify } from 'node:util';
-
-const execFileAsync = promisify( execFile );
-const maxDiffLength = 60_000;
-const e2eSpecPattern = /^test\/e2e\/specs\/.+\.spec\.(?:js|ts)$/;
-
-async function git( workspace, args ) {
-	const { stdout } = await execFileAsync( 'git', args, {
-		cwd: workspace,
-		maxBuffer: 10 * 1024 * 1024,
-	} );
-	return stdout;
-}
-
-async function collectArtifact( workspace ) {
-	// Intent-to-add makes untracked files visible to `git diff` without
-	// staging their contents in the disposable workspace.
-	await git( workspace, [ 'add', '--intent-to-add', '--all' ] );
-	const names = await git( workspace, [
-		'diff',
-		'--name-only',
-		'-z',
-		'HEAD',
-		'--',
-	] );
-	const changedFiles = names.split( '\0' ).filter( Boolean ).sort();
-	let diff = await git( workspace, [
-		'diff',
-		'--no-ext-diff',
-		'--unified=20',
-		'HEAD',
-		'--',
-	] );
-
-	const truncated = diff.length > maxDiffLength;
-	if ( truncated ) {
-		diff = `${ diff.slice( 0, maxDiffLength ) }\n[diff truncated]`;
-	}
-
-	return { changedFiles, diff, truncated };
-}
-
-function parseArtifact( output ) {
-	const result = typeof output === 'string' ? JSON.parse( output ) : output;
-	if (
-		! result?.artifact ||
-		! Array.isArray( result.artifact.changedFiles )
-	) {
-		throw new Error(
-			'The transformed output does not contain a workspace artifact.'
-		);
-	}
-	return result.artifact;
-}
+import {
+	getAddedDiffLines,
+	parseWorkspaceArtifact,
+} from '../../lib/workspace-artifact.mjs';
 
 function grade( pass, reason ) {
 	return { pass, score: pass ? 1 : 0, reason };
 }
 
-function addedLines( diff ) {
-	return diff
-		.split( '\n' )
-		.filter(
-			( line ) => line.startsWith( '+' ) && ! line.startsWith( '+++' )
-		)
-		.map( ( line ) => line.slice( 1 ) )
-		.join( '\n' );
-}
-
-export async function includeWorkspaceArtifact( output, context ) {
-	const workspace = context.vars.__workspace;
-	if ( typeof workspace !== 'string' ) {
-		throw new Error( 'The evaluation workspace was not provided.' );
-	}
-
-	return JSON.stringify(
-		{
-			response: output,
-			artifact: await collectArtifact( workspace ),
-		},
-		null,
-		2
-	);
-}
+const e2eSpecPattern = /^test\/e2e\/specs\/.+\.spec\.(?:js|ts)$/;
 
 export function changedE2eSpec( output ) {
-	const { changedFiles } = parseArtifact( output );
+	const { changedFiles } = parseWorkspaceArtifact( output );
 	const specs = changedFiles.filter( ( file ) =>
 		e2eSpecPattern.test( file )
 	);
@@ -101,7 +26,7 @@ export function changedE2eSpec( output ) {
 }
 
 export function changedOnlyE2eTests( output ) {
-	const { changedFiles } = parseArtifact( output );
+	const { changedFiles } = parseWorkspaceArtifact( output );
 	const unrelated = changedFiles.filter(
 		( file ) => ! file.startsWith( 'test/e2e/' )
 	);
@@ -114,7 +39,9 @@ export function changedOnlyE2eTests( output ) {
 }
 
 export function usedAccessibleParagraphLocator( output ) {
-	const additions = addedLines( parseArtifact( output ).diff );
+	const additions = getAddedDiffLines(
+		parseWorkspaceArtifact( output ).diff
+	);
 	const hasLocator =
 		/editor\s*\.\s*canvas\s*\.\s*getByRole\s*\(\s*(['"])document\1\s*,\s*\{[\s\S]{0,300}?name\s*:\s*(['"])Block:\s*Paragraph\2[\s\S]{0,100}?\}\s*\)/.test(
 			additions
@@ -128,7 +55,9 @@ export function usedAccessibleParagraphLocator( output ) {
 }
 
 export function enteredLiteralAmpersand( output ) {
-	const additions = addedLines( parseArtifact( output ).diff );
+	const additions = getAddedDiffLines(
+		parseWorkspaceArtifact( output ).diff
+	);
 	const hasInput =
 		/(?:fill|type|pressSequentially|insertText)\s*\(\s*(['"`])&\1/.test(
 			additions
@@ -142,7 +71,9 @@ export function enteredLiteralAmpersand( output ) {
 }
 
 export function assertedUnencodedAmpersand( output ) {
-	const additions = addedLines( parseArtifact( output ).diff );
+	const additions = getAddedDiffLines(
+		parseWorkspaceArtifact( output ).diff
+	);
 	const assertsRawAmpersand =
 		/(?:toBe|toEqual|toContain|toHaveText|toMatch|toMatchObject)\s*\([\s\S]{0,800}?(['"`])&(?!amp;)\1/.test(
 			additions
