+++ test/ai-development/evals/suites/testing-skill-routing/artifact-grader.test.mjs
@@ -2,23 +2,15 @@
  * Tests for this suite's artifact grader.
  */
 import assert from 'node:assert/strict';
-import { execFile } from 'node:child_process';
-import fs from 'node:fs/promises';
-import os from 'node:os';
-import path from 'node:path';
 import test from 'node:test';
-import { promisify } from 'node:util';
 
 import {
 	assertedUnencodedAmpersand,
 	changedE2eSpec,
 	changedOnlyE2eTests,
 	enteredLiteralAmpersand,
-	includeWorkspaceArtifact,
 	usedAccessibleParagraphLocator,
 } from './artifact-grader.mjs';
-
-const execFileAsync = promisify( execFile );
 
 function output( changedFiles, additions ) {
 	return JSON.stringify( {
@@ -95,46 +87,4 @@ test( 'recognizes a block attribute object assertion', () => {
 	);
 
 	assert.equal( assertedUnencodedAmpersand( result ).pass, true );
-} );
-
-test( 'captures an untracked file from the temporary workspace', async () => {
-	const workspace = await fs.mkdtemp(
-		path.join( os.tmpdir(), 'artifact-grader-test-' )
-	);
-	const spec = 'test/e2e/specs/editor/example.spec.js';
-
-	try {
-		await execFileAsync( 'git', [ 'init', '--quiet' ], { cwd: workspace } );
-		await execFileAsync(
-			'git',
-			[
-				'-c',
-				'user.name=Test',
-				'-c',
-				'user.email=test@example.com',
-				'commit',
-				'--allow-empty',
-				'--quiet',
-				'--message=Initial',
-			],
-			{ cwd: workspace }
-		);
-		await fs.mkdir( path.join( workspace, path.dirname( spec ) ), {
-			recursive: true,
-		} );
-		await fs.writeFile(
-			path.join( workspace, spec ),
-			"await paragraph.fill( '&' );\n"
-		);
-
-		const transformed = await includeWorkspaceArtifact( 'Done.', {
-			vars: { __workspace: workspace },
-		} );
-		const result = JSON.parse( transformed );
-
-		assert.deepEqual( result.artifact.changedFiles, [ spec ] );
-		assert.match( result.artifact.diff, /paragraph\.fill/ );
-	} finally {
-		await fs.rm( workspace, { recursive: true, force: true } );
-	}
 } );
